--- docs/cog_jsonscanner.rst
@@ -1,66 +1,0 @@
-.. _jsonscanner:
-
-===========
-JsonScanner
-===========
-
-This is the cog guide for the jsonscanner cog. You will
-find detailed docs about usage and commands.
-
-Throughout this documentation, ``[p]`` is considered as your prefix.
-
-------------
-Installation
-------------
-
-Let's firstly add my repository if you haven't already:
-
-* :code:`[p]repo add kreusada https://github.com/Kreusada/Kreusada-Cogs`
-
-Next, let's download the cog from the repo:
-
-* :code:`[p]cog install kreusada jsonscanner`
-
-Finally, you can see my end user data statements, cog requirements, and other cog information by using:
-
-* :code:`[p]cog info kreusada jsonscanner`
-
-.. _jsonscanner-usage:
-
------
-Usage
------
-
-An easy and quick tool to validate json.
-
-
-.. _jsonscanner-commands:
-
---------
-Commands
---------
-
-.. _jsonscanner-command-jsonscan:
-
-^^^^^^^^
-jsonscan
-^^^^^^^^
-
-**Syntax**
-
-.. code-block:: none
-
-    [p]jsonscan
-
-**Description**
-
-Scan json to see if its correct.
-
-Your next message will be used to as the json to scan.
-You can also upload a JSON file.
-
----------------
-Receive Support
----------------
-
-Feel free to ping me at the `Red Cog Support Server <https://discord.gg/GET4DVk>`_ in :code:`#support_kreusada-cogs`.
