+++ filev/__init__.py
@@ -1,0 +1,4 @@
+from .filev import Filev, __red_end_user_data_statement__
+
+def setup(bot):
+    bot.add_cog(Filev(bot))
