--- jsonscanner/__init__.py
@@ -1,4 +1,0 @@
-from .scanner import JsonScanner
-
-def setup(bot):
-    bot.add_cog(JsonScanner(bot))
