--- yamlscanner/__init__.py
@@ -1,4 +1,0 @@
-from .scanner import YamlScanner
-
-def setup(bot):
-    bot.add_cog(YamlScanner(bot))
