--- yamlscanner/info.json
@@ -1,11 +1,0 @@
-{
-    "author": ["Kreusada"],
-    "description": "An easy and quick tool to validate yaml.",
-    "install_msg": "Thanks for installing, have fun. Please refer to my docs if you need any help: https://kreusadacogs.readthedocs.io/en/latest/cog_yamlscanner.html",
-    "short": "An easy and quick tool to validate yaml.",
-    "name": "YamlScanner",
-    "tags": ["YAML", "Scanner", "Validator", "File"],
-    "requirements": [],
-    "type": "COG",
-    "end_user_data_statement": "This cog does not persistently store data or metadata about users."
-}
